move layout to components
