+++ packages/docz-theme/src/components/Layout/index.js
@@ -1,0 +1,1 @@
+export { Layout } from "./Layout";
