+++ packages/docz-theme/src/components/Sidebar/Sidebar.jsx
@@ -1,0 +1,6 @@
+/** @jsx jsx */
+import { Box, jsx } from "theme-ui";
+
+export const Sidebar = () => {
+  return <Box>Sidebar</Box>;
+};
