+++ packages/docz-theme/src/components/Sidebar/index.js
@@ -1,0 +1,1 @@
+export { Sidebar } from "./Sidebar";
